pyproject.toml
[project]
name = "tg-pay-bot"
version = "0.1.0"
description = "Async Telegram bot with user registration, payments and referral system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.4.0",
    "aiohttp-socks>=0.8.0",
    "sqlmodel>=0.0.16",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"